refactor: modularize FunctionPlotPage
